# contract/openapi.yaml
openapi: 3.0.3
info:
  title: Projects API
  version: 1.0.0
  description: |
    唯一 API 真相。Frontend 與 Backend 的型別皆由此檔生成。
    任何修改必須由 Architect 執行,並附帶一份 change-requests/CR-*.md。

servers:
  - url: http://localhost:3000
    description: local
  - url: https://staging.example.com
    description: staging

paths:
  /projects:
    get:
      operationId: listProjects
      summary: 列出目前租戶的所有專案
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ProjectStatus'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
    post:
      operationId: createProject
      summary: 建立專案
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: 驗證失敗
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /projects/{projectId}:
    get:
      operationId: getProject
      summary: 取得單一專案
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: 找不到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    ProjectStatus:
      type: string
      enum: [active, archived]

    Project:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          $ref: '#/components/schemas/ProjectStatus'
        createdAt:
          type: string
          format: date-time

    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string